Refactored OLED code
