+++ keyboards/splitkb/kyria/keymaps/bbaserdem/config.h
@@ -25,9 +25,4 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     #ifdef OLED_ENABLE
     #define OLED_DISPLAY_128X64
     #endif
-
-    // If you are using an Elite C rev3 on the slave side, uncomment the lines below:
-    // However; this causes the halves to be unresponsive
-    // #define SPLIT_USB_DETECT
-    // #define NO_USB_STARTUP_CHECK
 #endif
