+++ keyboards/splitkb/kyria/keymaps/bbaserdem/keymap.c
@@ -89,12 +89,24 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     )
 };
 
-// Kyria specific code
-#ifdef KEYBOARD_splitkb_kyria_rev1
-    // Rotate the display on kyria
-    #ifdef OLED_ENABLE
-        oled_rotation_t oled_init_user(oled_rotation_t rotation) {
-	    return OLED_ROTATION_180;
-        }
-    #endif //OLED_ENABLE
-#endif //KEYBOARD_kyria_rev1
+#ifdef KEYBOARD_splitkb_kyria_rev1      // Kyria specific code
+#ifdef OLED_ENABLE                          // Kyria's oled screen
+// Rotate the display on kyria
+oled_rotation_t oled_init_user(oled_rotation_t rotation) {
+    return OLED_ROTATION_180;
+}
+// QMK logo screen renderer; will render 126x32 image; but on the default font.
+static void render_qmk_logo(void) {
+    static const char PROGMEM qmk_logo[] = {
+        0x80,0x81,0x82,0x83,0x84,0x85,0x86,0x87,0x88,0x89,0x8a,0x8b,0x8c,0x8d,0x8e,0x8f,0x90,0x91,0x92,0x93,0x94,
+        0xa0,0xa1,0xa2,0xa3,0xa4,0xa5,0xa6,0xa7,0xa8,0xa9,0xaa,0xab,0xac,0xad,0xae,0xaf,0xb0,0xb1,0xb2,0xb3,0xb4,
+        0xc0,0xc1,0xc2,0xc3,0xc4,0xc5,0xc6,0xc7,0xc8,0xc9,0xca,0xcb,0xcc,0xcd,0xce,0xcf,0xd0,0xd1,0xd2,0xd3,0xd4,0};
+    oled_write_P(qmk_logo, false);
+}
+// For kyria; print logo, and allow default status printing
+bool oled_task_keymap(void) {
+    render_qmk_logo();
+    return true;
+}
+#endif //OLED_ENABLE
+#endif //KEYBOARD_splitkb_kyria_rev1
