+++ layouts/community/split_3x6_3/bbaserdem/rules.mk
@@ -40,4 +40,5 @@ ifneq (,$(findstring crkbd/rev1,$(KEYBOARD)))
     # BOOTMAGIC_ENABLE = lite
     # LTO_ENABLE = yes
     # PIMORONI_TRACKBALL = no
+    SRC += bb-crkbdarm-oled.c
 endif
