+++ users/bbaserdem/bb-encoder.c
@@ -25,8 +25,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * Modes can be reset using OS click
  * Some modes are only accessible through some keymap layers
  */
-//encoder_state_t encoder_state[NUMBER_OF_ENCODERS];
+
+// Initialize the encoder state variable
 encoder_state_t encoder_state[2];
+
 // Default state for the encoders
 void reset_encoder_state(void) {
 //    for (int i = 0; i < NUMBER_OF_ENCODERS; i++) {
@@ -40,181 +42,6 @@ void reset_encoder_state(void) {
 // Initialize the encoder state beginning
 void keyboard_post_init_encoder(void) {
     reset_encoder_state();
-}
-
-// Oled string printing for given layer and index
-void encoder_state_string_long(uint8_t index, uint8_t layer, char* buffer) {
-    // Get the layer straight from the main function
-    switch (layer) {
-        // If RGB control mode is enabled
-        #ifdef RGB_MATRIX_ENABLE
-        case _MEDI:
-            switch (encoder_state[index].rgb) {
-                case 0:
-                    strcpy(buffer, "ani mode");
-                    break;
-                case 1:
-                    strcpy(buffer, "hue     ");
-                    break;
-                case 2:
-                    strcpy(buffer, "saturat.");
-                    break;
-                case 3:
-                    strcpy(buffer, "bright. ");
-                    break;
-                case 4:
-                    strcpy(buffer, "ani. spd");
-                    break;
-                default:
-                    strcpy(buffer, " -N/A-  ");
-                    break;
-            }
-            break;
-        #endif
-        // If pointer control is enabled
-        #ifdef MOUSEKEY_ENABLE
-        case _MOUS:
-            switch (encoder_state[index].point) {
-                case 0:
-                    strcpy(buffer, "Lateral ");
-                    break;
-                case 1:
-                    strcpy(buffer, "Vertical");
-                    break;
-                case 2:
-                    strcpy(buffer, "Scr. Ver");
-                    break;
-                case 3:
-                    strcpy(buffer, "Scr. Lat");
-                    break;
-                default:
-                    strcpy(buffer, " -N/A-  ");
-                    break;
-            }
-            break;
-        #endif
-        default:
-            switch (encoder_state[index].base) {
-                case 0:
-                    strcpy(buffer, "Volume  ");
-                    break;
-                case 1:
-                    strcpy(buffer, "Song    ");
-                    break;
-                case 2:
-                    strcpy(buffer, "Sink    ");
-                    break;
-                case 3:
-                    strcpy(buffer, "Src. Vol");
-                    break;
-                case 4:
-                    strcpy(buffer, "Source  ");
-                    break;
-                case 5:
-                    strcpy(buffer, "Arrow LR");
-                    break;
-                case 6:
-                    strcpy(buffer, "Arrow UD");
-                    break;
-                case 7:
-                    strcpy(buffer, "Page U/D");
-                    break;
-                case 8:
-                    strcpy(buffer, "Erase   ");
-                    break;
-                default:
-                    strcpy(buffer, " -N/A-  ");
-                    break;
-            }
-            break;
-    }
-}
-
-void encoder_state_string(uint8_t index, uint8_t layer, char* buffer) {
-    // Get the layer straight from the main function
-    switch (layer) {
-        // If RGB control mode is enabled
-        #ifdef RGB_MATRIX_ENABLE
-        case _MEDI:
-            switch (encoder_state[index].rgb) {
-                case 0:
-                    strcpy(buffer, " mode");
-                    break;
-                case 1:
-                    strcpy(buffer, " hue ");
-                    break;
-                case 2:
-                    strcpy(buffer, "satur");
-                    break;
-                case 3:
-                    strcpy(buffer, "value");
-                    break;
-                case 4:
-                    strcpy(buffer, "speed");
-                    break;
-                default:
-                    strcpy(buffer, "!oob!");
-                    break;
-            }
-            break;
-        #endif
-        // If pointer control is enabled
-        #ifdef MOUSEKEY_ENABLE
-        case _MOUS:
-            switch (encoder_state[index].point) {
-                case 0:
-                    strcpy(buffer, "m.lat");
-                    break;
-                case 1:
-                    strcpy(buffer, "m.ver");
-                    break;
-                case 2:
-                    strcpy(buffer, "s.ver");
-                    break;
-                case 3:
-                    strcpy(buffer, "s.lat");
-                    break;
-                default:
-                    strcpy(buffer, "!oob!");
-                    break;
-            }
-            break;
-        #endif
-        default:
-            switch (encoder_state[index].base) {
-                case 0:
-                    strcpy(buffer, " volm");
-                    break;
-                case 1:
-                    strcpy(buffer, " song");
-                    break;
-                case 2:
-                    strcpy(buffer, " sink");
-                    break;
-                case 3:
-                    strcpy(buffer, "s.vol");
-                    break;
-                case 4:
-                    strcpy(buffer, " src ");
-                    break;
-                case 5:
-                    strcpy(buffer, " L/R ");
-                    break;
-                case 6:
-                    strcpy(buffer, " U/D ");
-                    break;
-                case 7:
-                    strcpy(buffer, "pgU/D");
-                    break;
-                case 8:
-                    strcpy(buffer, " del ");
-                    break;
-                default:
-                    strcpy(buffer, "!oob!");
-                    break;
-            }
-            break;
-    }
 }
 
 // Encoder scroll functionality
