+++ users/bbaserdem/bb-oled.c
@@ -12,383 +12,199 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 #include "bb-oled.h"
+#include <string.h>
+// Allow access to encoder_state
 #ifdef ENCODER_ENABLE
-    #include "bb-encoder.h"
-#endif
-/* OLED ENCODER
- * This contains my general oled code
- * Currently I have two setups with OLEDs
- * - Kyria; with 128x64 but with low memory (elite C)
- * - Protorne; wh 128x32 but with excellent memory
- * This file differentiates only between those two
+#include "bb-encoder.h"
+#endif // ENCODER_ENABLE
+
+/* OLED
+ * This contains general purpose oled code
  */ 
 
-// Following line imports reading of RGB settings
-#if defined RGB_MATRIX_ENABLE
-extern rgb_config_t rgb_matrix_config;
-#endif
+#ifdef ENCODER_ENABLE
 
-#if defined(KEYBOARD_crkbd_rev1)
-// These are my images for keyboard layout
-// Naming scheme is <layer>_left/right{_keyboardLayout}{_modifiers}
-static const char PROGMEM base0_L_dvor_nomod[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x04,0x08,0x10,0x00,0x00,0x00,
-        0x00,0x0e,0x00,0x00,0x00,0x00,0x60,0xe0,0x00,0x00,0x00,0x00,0xc0,0xc0,
-        0x00,0x00,0x00,0xf8,0x90,0x88,0x88,0x70,0x00,0x78,0x80,0x80,0x40,0xf8
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x00,0x00,0x80,0x60,0x18,0x00,0x00,
-        0x40,0x40,0x40,0x80,0x00,0x80,0x41,0x40,0x40,0x80,0x00,0x80,0x40,0x40,
-        0x40,0x80,0x00,0xc3,0x00,0x00,0x00,0xc0,0x00,0x00,0x42,0xd2,0x02,0x01
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0xd8,0x06,0x01,0x00,0x00,0x00,0x02,
-        0x05,0x05,0x05,0x07,0x00,0x03,0x04,0x04,0x04,0x03,0x00,0x03,0x05,0x05,
-        0x85,0x05,0x00,0xc3,0x04,0x04,0x02,0x07,0x00,0x00,0x00,0x07,0x04,0x00
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x00,0x03,0x0c,0x30,0xc0,0x00,0x00,
-        0xb6,0x76,0x00,0x00,0x00,0x1c,0x22,0x22,0x12,0xfe,0x00,0x00,0x80,0x82,
-        0x7e,0x00,0x00,0x3f,0x08,0x0c,0x12,0x20,0x00,0x22,0x14,0x08,0x14,0x22}};
-static const char PROGMEM base0_L_dvor_shift[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x20,0x10,0x20,0x40,0x20,0x00,0x00,
-        0x07,0x00,0x07,0x00,0x00,0x10,0x28,0x44,0x82,0x00,0x00,0x82,0x44,0x28,
-        0x10,0x00,0x00,0xfe,0x12,0x12,0x12,0x0c,0x00,0x06,0x08,0xf0,0x08,0x06
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x00,0x00,0xf0,0x00,0x00,0x00,0xe0,
-        0x90,0x90,0x90,0xe0,0x00,0xe0,0x10,0x10,0x10,0xe0,0x00,0xf0,0x90,0x90,
-        0x90,0x10,0x00,0xf0,0x00,0x00,0x00,0xf0,0x00,0x00,0x10,0xf0,0x10,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0x00,0x80,0x87,0x80,0x00,0x00,0x07,
-        0x00,0x00,0x00,0x07,0x00,0x03,0x84,0x84,0x84,0x03,0x00,0x07,0x04,0x04,
-        0x04,0x84,0x00,0x83,0x04,0x04,0x04,0x83,0x00,0x80,0x04,0x07,0x04,0x80
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x01,0x00,0x28,0x04,0x03,0x00,0x00,
-        0x36,0x36,0x00,0x00,0x00,0x1f,0x20,0x28,0x10,0x6f,0x00,0x18,0x20,0x20,
-        0x20,0x1f,0x00,0x3f,0x04,0x0a,0x11,0x20,0x00,0x31,0x0a,0x04,0x0a,0x31}};
-//static const char PROGMEM base0_R_dvor_nomod[][4] = {};
-//static const char PROGMEM base0_R_dvor_shift[][4] = {};
-static const char PROGMEM char1_L[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0xe2,0xae,0x42,0x00,0x00,0x20,0x44,0x40,0x4c,0x20,0x00,0x12,0xa4,0x88,
-        0xa4,0x12,0x00,0x08,0xf8,0x08,0xf8,0x08,0x00,0x78,0x80,0x80,0x80,0x78
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,
-        0x41,0x40,0x41,0x80,0x00,0x80,0x40,0x40,0x40,0x80,0x00,0x81,0x40,0x40,
-        0x40,0x81,0x00,0xc0,0x00,0xc0,0x00,0xc0,0x00,0x00,0xc0,0x00,0x00,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x83,
-        0xc4,0x84,0xc2,0x07,0x00,0x03,0x04,0x04,0x04,0x03,0x00,0x02,0x85,0x05,
-        0x04,0x82,0x00,0x03,0x04,0x1f,0x04,0x03,0x00,0x80,0x83,0x84,0x82,0x80
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7f,
-        0xc4,0x44,0xef,0x3b,0x00,0x3e,0x49,0x49,0x49,0x3e,0x00,0x02,0x7f,0x09,
-        0x09,0x78,0x00,0x3e,0x08,0x08,0x14,0x22,0x00,0x20,0x24,0x24,0x24,0x20}};
-//static const char PROGMEM char1_R[][4] = {};
-static const char PROGMEM game2_L[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7c,
-        0x82,0xa2,0x42,0xbc,0x00,0x3e,0xc0,0x30,0xc0,0x3e,0x00,0xfe,0x92,0x92,
-        0x92,0x82,0x00,0xfe,0x12,0x12,0x32,0xcc,0x00,0x02,0x02,0xfe,0x02,0x02
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x40,0x40,0x40,0x40,0xe0,0x00,0xe0,
-        0x90,0x90,0x90,0xe1,0x00,0x60,0x90,0x90,0x90,0x20,0x00,0xf0,0x10,0x10,
-        0x10,0xe0,0x00,0xf0,0x90,0x90,0x90,0x10,0x00,0xe0,0x10,0x90,0x90,0xa0
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0x07,0x02,0x02,0x02,0x02,0x00,0x87,
-        0x80,0x80,0x80,0x87,0x00,0x82,0x04,0x04,0x04,0x83,0x00,0x07,0x84,0x84,
-        0x84,0x03,0x00,0x87,0x00,0x00,0x00,0x80,0x00,0x83,0x84,0x84,0x84,0x03
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x08,0x3c,0x02,0x3c,0x08,0x00,0x30,
-        0x28,0x24,0x22,0x21,0x00,0x31,0x0a,0x04,0x0a,0x31,0x00,0x1f,0x20,0x20,
-        0x20,0x11,0x00,0x03,0x0c,0x30,0x0c,0x03,0x00,0x3f,0x24,0x24,0x24,0x1b}};
-//static const char PROGMEM medi3_R[][4] = {};
-//static const char PROGMEM navi4_R[][4] = {};
-//static const char PROGMEM symb5_R_dvor[][4] = {};
-static const char PROGMEM numb6_L_dvor_nomod[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,
-        0x50,0x88,0x04,0x00,0x00,0x04,0x04,0xc4,0x34,0x0c,0x00,0xd8,0x24,0x24,
-        0x24,0xd8,0x00,0x18,0x24,0x24,0x24,0xf8,0x00,0xf8,0x44,0x24,0x14,0xf8
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0x00,0x00,0x01,0x00,0x00,0x00,0x80,0x41,0xe0,0x00,0x00,0xe0,0xa1,0xa1,
-        0xa1,0x20,0x00,0xc0,0xa1,0xa1,0xa1,0x00,0x00,0x00,0x01,0x01,0x01,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,
-        0x08,0x08,0x08,0x08,0x00,0x03,0x02,0x02,0x0f,0x02,0x00,0x04,0x08,0x08,
-        0x08,0x07,0x00,0x07,0x08,0x08,0x08,0x07,0x00,0x01,0x01,0x01,0x01,0x00
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x41,
-        0x22,0x14,0x08,0x00,0x00,0x00,0x04,0x02,0x7f,0x00,0x00,0x42,0x61,0x51,
-        0x49,0x46,0x00,0x22,0x41,0x49,0x49,0x36,0x00,0x5e,0x61,0x01,0x61,0x5e}};
-static const char PROGMEM numb6_L_dvor_shift[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0x00,0x00,0x00,0x00,0x00,0xd8,0x24,0x54,0x88,0x40,0x00,0x50,0x20,0xf8,
-        0x20,0x50,0x00,0x00,0x38,0xc6,0x01,0x00,0x00,0x00,0x01,0xc6,0x38,0x00
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0x00,0x00,0x00,0x00,0x00,0x60,0x91,0xf9,0x90,0x21,0x00,0x20,0x50,0x20,
-        0x80,0x60,0x00,0x40,0x20,0x10,0x21,0x40,0x00,0x00,0x01,0x00,0x00,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0x00,0x00,0x00,0x00,0x00,0x02,0x04,0x8f,0x04,0x03,0x00,0x0c,0x02,0x09,
-        0x14,0x08,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-        0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x5f,0x00,0x00,0x00,0x3e,0x41,0x5d,
-        0x55,0x5e,0x00,0x14,0x7f,0x14,0x7f,0x14,0x00,0x00,0x00,0x00,0x00,0x00}};
-static const char PROGMEM func7_L[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0xff,0x09,0xe9,0xc9,0x01,0x00,0xe0,0x00,
-        0x10,0x08,0xfc,0x00,0x00,0x08,0x84,0x44,0x24,0x18,0x00,0x88,0x04,0x24,
-        0x24,0xd8,0x00,0x60,0x50,0x48,0xfc,0x40,0x00,0x70,0x80,0x9c,0x80,0x70
-    },{ 0x00,0x00,0xce,0x9c,0x00,0xf0,0x00,0x00,0x9f,0x00,0x03,0x0c,0x1f,0x00,
-        0xe0,0xa0,0xa1,0x20,0x00,0xc1,0xa1,0xa1,0xa1,0x01,0x00,0x20,0x21,0x21,
-        0xa1,0x60,0x00,0xc0,0x20,0x20,0x21,0xc0,0x00,0x40,0xe0,0x50,0x40,0x80
-    },{ 0x00,0x00,0x39,0x73,0x00,0xef,0x02,0x83,0xe4,0x00,0x0e,0x15,0x16,0x00,
-        0x04,0x08,0x08,0x07,0x00,0x07,0x48,0xc8,0x08,0x07,0x00,0x00,0x40,0xce,
-        0x01,0x00,0x00,0x06,0x49,0xc9,0x09,0x06,0x00,0x06,0x08,0x49,0x08,0x07
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x05,0x04,0x03,0x90,0xa8,0xa8,0x40,0x06,
-        0x49,0x49,0x49,0x3e,0x00,0x7c,0xa2,0x93,0x8a,0x7c,0x00,0x00,0x08,0x07,
-        0xfe,0x00,0x00,0x84,0xc2,0xa3,0x92,0x8c,0x00,0x70,0x38,0x0d,0x38,0x70}};
-static const char PROGMEM mous8_L[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x3c,0x84,0xbc,0x84,0x38,0x00,0x00,0x00,0x10,
-        0x28,0x92,0xc5,0x82,0x00,0xfe,0x1f,0x1f,0x01,0xfe,0x00,0xfe,0x01,0x19,
-        0x01,0xfe,0x00,0xfe,0x01,0x1f,0x1f,0xfe,0x00,0x10,0x38,0x92,0xc7,0x82
-    },{ 0x00,0x00,0xce,0x9c,0x00,0xe7,0x08,0x08,0x08,0xe7,0x00,0x00,0x00,0x00,
-        0x00,0x80,0x01,0x00,0x00,0x01,0x01,0x81,0x01,0x01,0x00,0x01,0x01,0x81,
-        0x01,0x01,0x00,0x01,0x01,0x81,0x01,0x01,0x00,0x60,0x50,0xf8,0x41,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x91,0xaa,0xaa,0xa9,0x43,0x00,0x00,0x00,0x02,
-        0x07,0xca,0x02,0x02,0x00,0x02,0x01,0xcf,0x01,0x02,0x00,0x02,0x04,0x0f,
-        0x04,0x02,0x00,0x02,0x02,0xca,0x07,0x02,0x00,0xdc,0x58,0x5c,0x18,0x1f
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x1c,0x2a,0x2a,0x2a,0x2c,0x00,0x00,0x00,0x0c,
-        0x33,0xc0,0x00,0x00,0x00,0x0c,0x03,0x00,0x03,0x0c,0x00,0x0c,0x30,0xc0,
-        0x30,0x0c,0x00,0x00,0x00,0xc0,0x33,0x0c,0x00,0xe5,0xc9,0xe6,0xc0,0xf8}};
-static const char PROGMEM musi9_L[4][42] = {
-    {   0x00,0x00,0x73,0xe7,0x00,0x00,0x00,0x00,0x00,0xfe,0x04,0x18,0x04,0xfe,
-        0x00,0x00,0xfe,0x00,0x00,0x00,0xfe,0x00,0x8c,0x12,0x12,0x12,0xe4,0x00,
-        0x00,0x02,0xfe,0x02,0x00,0x00,0xfc,0x02,0x02,0x02,0x84,0x00,0x00,0x00
-    },{ 0x00,0x00,0xce,0x9c,0x00,0x00,0xc0,0xe0,0xf0,0xf9,0xf8,0xf8,0xf8,0xf9,
-        0xf0,0xe0,0xc0,0x01,0x01,0x01,0x00,0xf0,0xf0,0x01,0x01,0xf1,0xf0,0x00,
-        0x00,0x01,0x01,0x01,0x00,0x00,0xf8,0xf1,0xe1,0xc1,0x80,0x00,0x00,0x00
-    },{ 0x00,0x00,0x39,0x73,0x00,0x00,0x07,0x0f,0x1f,0x3f,0x3f,0x3f,0x3f,0x3f,
-        0x1f,0x0f,0x07,0x00,0x00,0x00,0x00,0x1f,0x1f,0x00,0x00,0x1f,0x1f,0x00,
-        0x00,0x00,0x00,0x00,0x00,0x00,0x3f,0x1f,0x0f,0x07,0x03,0x01,0x00,0x00
-    },{ 0x00,0x00,0xe7,0xce,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xf8,0x08,
-        0xf8,0x08,0xf0,0x00,0x70,0x88,0x88,0x88,0x70,0x00,0x70,0x88,0x88,0x48,
-        0xff,0x00,0x70,0xa8,0xa8,0xa8,0xb0,0x00,0x00,0x00,0x00,0x00,0x00,0x00}};
-static const char PROGMEM musi9_R[4][42] = {
-    {   0x00,0x00,0x00,0xfe,0x04,0x18,0x04,0xfe,0x00,0x00,0xfe,0x00,0x00,0x00,
-        0xfe,0x00,0x8c,0x12,0x12,0x12,0xe4,0x00,0x00,0x02,0xfe,0x02,0x00,0x00,
-        0xfc,0x02,0x02,0x02,0x84,0x00,0x00,0x00,0x00,0x00,0xe7,0x73,0x00,0x00
-    },{ 0x00,0x00,0x00,0x81,0x60,0x10,0x60,0x81,0x60,0x10,0x00,0x01,0x01,0x01,
-        0xe0,0xe0,0x60,0x61,0x61,0x61,0x60,0xf8,0x70,0x21,0x01,0x01,0xc0,0xe0,
-        0x60,0x01,0xf9,0xf9,0x00,0x60,0xe0,0xc0,0x00,0x00,0x9c,0xce,0x00,0x00
-    },{ 0x18,0x24,0x4b,0x48,0x48,0x4c,0x4b,0x48,0x4b,0x2c,0x18,0x00,0x00,0x00,
-        0x4f,0xe7,0xf0,0x60,0x60,0x60,0x60,0x60,0x7e,0x7f,0x00,0x0f,0x3f,0x70,
-        0x60,0xc0,0xc1,0xc1,0xc0,0x60,0x70,0x3f,0x0f,0x00,0x73,0x39,0x00,0x00
-    },{ 0x00,0x00,0x00,0x00,0x00,0x00,0xf8,0x08,0xf8,0x08,0xf0,0x00,0x70,0x88,
-        0x88,0x88,0x71,0x00,0x70,0x88,0x88,0x48,0xff,0x00,0x70,0xa8,0xa8,0xa8,
-        0xb0,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0xce,0xe7,0x00,0x00}};
-static const char PROGMEM mod_meta[12] = {0x00,0x01,0x01,0x01,0x01,0x06,0x18,0x60,0x81,0x81,0x81,0x81};
-static const char PROGMEM mod_altL[12] = {0x00,0x80,0x80,0x80,0x40,0x20,0x10,0x08,0x84,0x95,0x8e,0x84};
-static const char PROGMEM mod_altR[12] = {0x00,0x86,0x8f,0x8f,0x46,0x20,0x10,0x08,0x84,0x95,0x8e,0x84};
-static const char PROGMEM mod_ctrl[12] = {0x00,0x00,0xbd,0x42,0xa5,0x99,0x99,0xa5,0x42,0xbd,0x00,0x00};
-static const char PROGMEM mod_shft[12] = {0x00,0x20,0x30,0x28,0xe4,0x02,0x01,0x02,0xe4,0x28,0x30,0x20};
-static const char PROGMEM rgb_enab[12] = {0x08,0x81,0x3c,0x42,0x99,0xbd,0xbd,0x99,0x42,0x3c,0x81,0x10};
-static const char PROGMEM rgb_disa[12] = {0x08,0x81,0x3c,0x42,0x81,0x81,0x81,0x81,0x42,0x3c,0x81,0x10};
-
-void render_modifiers(uint8_t row, uint8_t col, uint8_t mods) {
-    // Write modifiers starting from said column and row; does 14 characters
-    // Looks like:
-    //  Mods: <OS> <Alt/Gr> <Ctrl> <Shift>
-    oled_set_cursor(col, row);
-    oled_write("Mods: ", false);
-    // Meta
-    if (mods & MOD_MASK_GUI) {
-        oled_write_raw_P(mod_meta, 12);
-        oled_set_cursor(col + 8, row);
-    } else {
-        oled_write("  ", false);
-    }
-    // Alt(Gr)
-    if (mods & MOD_BIT(KC_RALT)) {
-        oled_write_raw_P(mod_altR, 12);
-        oled_set_cursor(col + 10, row);
-    } else if (mods & MOD_MASK_ALT) {
-        oled_write_raw_P(mod_altL, 12);
-        oled_set_cursor(col + 10, row);
-    } else {
-        oled_write("  ", false);
-    }
-    // Ctrl
-    if (mods & MOD_MASK_CTRL) {
-        oled_write_raw_P(mod_ctrl, 12);
-        oled_set_cursor(col + 12, row);
-    } else {
-        oled_write("  ", false);
-    }
-    // Shift
-    if (mods & MOD_MASK_SHIFT) {
-        oled_write_raw_P(mod_shft, 12);
-        oled_set_cursor(col + 14, row);
-    } else {
-        oled_write("  ", false);
-    }
-}
-
-void render_encoder(uint8_t row, uint8_t col, uint8_t index, uint8_t layer) {
-    // Renders the encoder state, 14 characters
-    static char encoder_temp9[9] = {0};
-
-    oled_set_cursor(col, row);
-    #if defined ENCODER_ENABLE
-    oled_write("Enc: ", false);
-    encoder_state_string_long(index, layer, encoder_temp9);
-    oled_write(encoder_temp9, false);
-    #else
-    oled_set_cursor(col, row);
-    oled_write("No enc. avail.", false);
-    #endif
-}
-
-void render_wpm(uint8_t row, uint8_t col) {
-    // Renders the WPM, 14 characters
-    static char wpm_temp4[4] = {0};
-
-    oled_set_cursor(col, row);
-    #if defined WPM_ENABLE
-    oled_write("WPM: ", false);
-    itoa(get_current_wpm(), wpm_temp4, 10);
-    oled_write(wpm_temp4, false);
-    oled_write("      ", false);
-    #else
-    oled_write("WPM not avail.", false);
-    #endif
-}
-
-void render_keymap(uint8_t row, uint8_t col) {
-    // Render the oled layout
-    oled_set_cursor(col, row);
-    oled_write("Layout: DVORAK", false);
-}
-
-void draw_image(uint8_t row, uint8_t col, const char image[4][42]) {
-    // Draw this image iteratively
-    for (int i = 0; i < 4; i++) {
-        oled_set_cursor(col, row + i);
-        oled_write_raw_P(image[i], 42);
-    }
-}
-
-void render_layout(uint8_t row, uint8_t col, uint8_t mods, bool isLeft) {
-    // Render the requested image on the given column and row 
-    // Don't grab highest layer; instead compare from top to bottom.
-    // Highest layer might be on the other side
-    if (isLeft) {
-        if (layer_state_cmp(layer_state, _MUSI)) {
-            draw_image(row, col, musi9_L);
-        } else if (layer_state_cmp(layer_state, _MOUS)) {
-            draw_image(row, col, mous8_L);
-        } else if (layer_state_cmp(layer_state, _FUNC)) {
-            draw_image(row, col, func7_L);
-        } else if (layer_state_cmp(layer_state, _NUMB)) {
-            if (mods & MOD_MASK_SHIFT) {
-                draw_image(row, col, numb6_L_dvor_shift);
-            } else {
-                draw_image(row, col, numb6_L_dvor_nomod);
+#if !defined KEYBOARD_splitkb_kyria_rev1 // Don't include in kyria; space
+void encoder_state_string8(uint8_t index, uint8_t layer, char* buffer) {
+    // Get the layer straight from the main function
+    switch (layer) {
+        // If RGB control mode is enabled
+        #ifdef RGB_MATRIX_ENABLE
+        case _MEDI:
+            switch (encoder_state[index].rgb) {
+                case 0:
+                    strcpy(buffer, "ani mode");
+                    break;
+                case 1:
+                    strcpy(buffer, "hue     ");
+                    break;
+                case 2:
+                    strcpy(buffer, "saturat.");
+                    break;
+                case 3:
+                    strcpy(buffer, "bright. ");
+                    break;
+                case 4:
+                    strcpy(buffer, "ani. spd");
+                    break;
+                default:
+                    strcpy(buffer, " -N/A-  ");
+                    break;
             }
-        } else if (layer_state_cmp(layer_state, _GAME)) {
-            draw_image(row, col, game2_L);
-        } else if (layer_state_cmp(layer_state, _CHAR)) {
-            draw_image(row, col, char1_L);
-        } else {
-            if (mods & MOD_MASK_SHIFT) {
-                draw_image(row, col, base0_L_dvor_shift);
-            } else {
-                draw_image(row, col, base0_L_dvor_nomod);
+            break;
+        #endif
+        // If pointer control is enabled
+        #ifdef MOUSEKEY_ENABLE
+        case _MOUS:
+            switch (encoder_state[index].point) {
+                case 0:
+                    strcpy(buffer, "Lateral ");
+                    break;
+                case 1:
+                    strcpy(buffer, "Vertical");
+                    break;
+                case 2:
+                    strcpy(buffer, "Scr. Ver");
+                    break;
+                case 3:
+                    strcpy(buffer, "Scr. Lat");
+                    break;
+                default:
+                    strcpy(buffer, " -N/A-  ");
+                    break;
             }
-        }
-    } else {
-        // WIP
-        draw_image(row, col, musi9_R);
+            break;
+        #endif
+        default:
+            switch (encoder_state[index].base) {
+                case 0:
+                    strcpy(buffer, "Volume  ");
+                    break;
+                case 1:
+                    strcpy(buffer, "Song    ");
+                    break;
+                case 2:
+                    strcpy(buffer, "Sink    ");
+                    break;
+                case 3:
+                    strcpy(buffer, "Src. Vol");
+                    break;
+                case 4:
+                    strcpy(buffer, "Source  ");
+                    break;
+                case 5:
+                    strcpy(buffer, "Arrow LR");
+                    break;
+                case 6:
+                    strcpy(buffer, "Arrow UD");
+                    break;
+                case 7:
+                    strcpy(buffer, "Page U/D");
+                    break;
+                case 8:
+                    strcpy(buffer, "Erase   ");
+                    break;
+                default:
+                    strcpy(buffer, " -N/A-  ");
+                    break;
+            }
+            break;
     }
 }
+#endif // !KEYBOARD_splitkb_kyria_rev1
 
-void render_rgb(uint8_t row, uint8_t col) {
-    // Render the RGB state on the given column and row 
-    static char rgb_temp4[4] = {0};
-    static char rgb_temp3[3] = {0};
-
-    oled_set_cursor(col, row);
-#if defined RGB_MATRIX_ENABLE
-    if (rgb_matrix_is_enabled()) {
-        oled_write_raw_P(rgb_enab, 12);
-    } else {
-        oled_write_raw_P(rgb_disa, 12);
+void encoder_state_string5(uint8_t index, uint8_t layer, char* buffer) {
+    // Get the layer straight from the main function
+    switch (layer) {
+        // If RGB control mode is enabled
+        #ifdef RGB_MATRIX_ENABLE
+        case _MEDI:
+            switch (encoder_state[index].rgb) {
+                case 0:
+                    strcpy(buffer, " mode");
+                    break;
+                case 1:
+                    strcpy(buffer, " hue ");
+                    break;
+                case 2:
+                    strcpy(buffer, "satur");
+                    break;
+                case 3:
+                    strcpy(buffer, "value");
+                    break;
+                case 4:
+                    strcpy(buffer, "speed");
+                    break;
+                default:
+                    strcpy(buffer, "!oob!");
+                    break;
+            }
+            break;
+        #endif
+        // If pointer control is enabled
+        #ifdef MOUSEKEY_ENABLE
+        case _MOUS:
+            switch (encoder_state[index].point) {
+                case 0:
+                    strcpy(buffer, "m.lat");
+                    break;
+                case 1:
+                    strcpy(buffer, "m.ver");
+                    break;
+                case 2:
+                    strcpy(buffer, "s.ver");
+                    break;
+                case 3:
+                    strcpy(buffer, "s.lat");
+                    break;
+                default:
+                    strcpy(buffer, "!oob!");
+                    break;
+            }
+            break;
+        #endif
+        default:
+            switch (encoder_state[index].base) {
+                case 0:
+                    strcpy(buffer, " volm");
+                    break;
+                case 1:
+                    strcpy(buffer, " song");
+                    break;
+                case 2:
+                    strcpy(buffer, " sink");
+                    break;
+                case 3:
+                    strcpy(buffer, "s.vol");
+                    break;
+                case 4:
+                    strcpy(buffer, " src ");
+                    break;
+                case 5:
+                    strcpy(buffer, " L/R ");
+                    break;
+                case 6:
+                    strcpy(buffer, " U/D ");
+                    break;
+                case 7:
+                    strcpy(buffer, "pgU/D");
+                    break;
+                case 8:
+                    strcpy(buffer, " del ");
+                    break;
+                default:
+                    strcpy(buffer, "!oob!");
+                    break;
+            }
+            break;
     }
-    oled_set_cursor(col + 2, row);
-    oled_write(" hue sat val", false);
-    oled_set_cursor(col, row + 1);
-    itoa(rgb_matrix_get_mode(), rgb_temp3, 10);
-    oled_write(rgb_temp3, false);
-    oled_write(" ", false);
-    itoa(rgb_matrix_get_hue(), rgb_temp4, 10);
-    oled_write(rgb_temp4, false);
-    oled_write(" ", false);
-    itoa(rgb_matrix_get_sat(), rgb_temp4, 10);
-    oled_write(rgb_temp4, false);
-    oled_write(" ", false);
-    itoa(rgb_matrix_get_val(), rgb_temp4, 10);
-    oled_write(rgb_temp4, false);
-#else
-    oled_write("-RGB disabled-", false);
-#endif
 }
+#endif // ENCODER_ENABLE
 
-void render_status_left(void) {
-    // Left side looks like this on the left half
-    // (Should be 14 characters per line)
-    //  Mods: <OS> <Alt/Gr> <Ctrl> <Shift>
-    //  Enc: <8 len str>
-    //  Wpm: <wpm here>
-    //  Layout: <dvorak/qwerty/tur. f>
-    // Right half is whatever layer image needs be
-    uint8_t this_mod =   get_mods();
-    uint8_t this_layer = get_highest_layer(layer_state);
-
-    // MODIFIERS
-    render_modifiers(0, 0, this_mod);
-
-    // Encoders
-    render_encoder(1, 0, 0, this_layer);
-    
-    // WPM text
-    render_wpm(2, 0);
-    
-    // Visual layout
-    render_keymap(3, 0);
-
-    // Draw the image after 14'th character
-    render_layout(0, 14, this_mod, true);
-}
-
-void render_status_right(void) {
-    // Right half is whatever layer image needs be on the left
-    // Right side looks like this on the right half
-    // (Should be after the 12'th character; max 14 characters per line)
-    //  Mods: <OS> <Alt/Gr> <Ctrl> <Shift>
-    //  Enc: <8 len str>
-    //  Wpm: <wpm here>
-    //  Layout: <dvorak/qwerty/tur. f>
-    uint8_t this_mod =   get_mods();
-    uint8_t this_layer = get_highest_layer(layer_state);
-
-    // Draw the image
-    render_layout(0, 0, this_mod, false);
-
-    // Modifiers
-    render_modifiers(0, 7, this_mod);
-
-    // Encoder
-    render_encoder(1, 7, 1, this_layer);
-
-    // RGB State; WIP
-    render_rgb(2, 7);
-}
-#endif
-
-#if defined(KEYBOARD_splitkb_kyria_rev1)
-static void render_qmk_logo(void) {
-  static const char PROGMEM qmk_logo[] = {
-    0x80,0x81,0x82,0x83,0x84,0x85,0x86,0x87,0x88,0x89,0x8a,0x8b,0x8c,0x8d,0x8e,0x8f,0x90,0x91,0x92,0x93,0x94,
-    0xa0,0xa1,0xa2,0xa3,0xa4,0xa5,0xa6,0xa7,0xa8,0xa9,0xaa,0xab,0xac,0xad,0xae,0xaf,0xb0,0xb1,0xb2,0xb3,0xb4,
-    0xc0,0xc1,0xc2,0xc3,0xc4,0xc5,0xc6,0xc7,0xc8,0xc9,0xca,0xcb,0xcc,0xcd,0xce,0xcf,0xd0,0xd1,0xd2,0xd3,0xd4,0};
-  oled_write_P(qmk_logo, false);
-}
 static void render_status(void) {
     // Function to print state information
     uint8_t this_layer = get_highest_layer(layer_state);
     uint8_t this_mod =   get_mods();
-    // Render qmk
-    render_qmk_logo();
     // Temporary variable for RGB
     #if defined RGB_MATRIX_ENABLE
     static char rgb_temp4[4] = {0};
@@ -443,16 +259,16 @@ static void render_status(void) {
 #if defined RGB_MATRIX_ENABLE
         case _MEDI:
             oled_write("m", false);
-            itoa(rgb_matrix_config.mode, rgb_temp4, 10);
+            itoa(rgb_matrix_get_mode(), rgb_temp4, 10);
             oled_write(rgb_temp4, false);
             oled_write(" h", false);
-            itoa(rgb_matrix_config.hsv.h, rgb_temp4, 10);
+            itoa(rgb_matrix_get_hue(), rgb_temp4, 10);
             oled_write(rgb_temp4, false);
             oled_write(" s", false);
-            itoa(rgb_matrix_config.hsv.s, rgb_temp4, 10);
+            itoa(rgb_matrix_get_sat(), rgb_temp4, 10);
             oled_write(rgb_temp4, false);
             oled_write(" v", false);
-            itoa(rgb_matrix_config.hsv.v, rgb_temp4, 10);
+            itoa(rgb_matrix_get_val(), rgb_temp4, 10);
             oled_write(rgb_temp4, false);
             oled_write("\n", false);
             break;
@@ -479,24 +295,23 @@ static void render_status(void) {
     // Line 4: Encoder states
 #ifdef ENCODER_ENABLE
     oled_write("EN0:", false);
-    encoder_state_string(0, this_layer, encoder_temp6);
+    encoder_state_string5(0, this_layer, encoder_temp6);
     oled_write(encoder_temp6, false);
     oled_write(" EN1:", false);
-    encoder_state_string(1, this_layer, encoder_temp6);
+    encoder_state_string5(1, this_layer, encoder_temp6);
     oled_write(encoder_temp6, false);
+#else
+    oled_write("-Encoder unavailable-", false);
 #endif
     oled_write("\n", false);
 }
-#endif
 
+// Allow default to be overwritten by keymap if they return false
+__attribute__ ((weak)) bool oled_task_keymap(void) {return true;}
+
+// Do sane defaults for regular oled rendering
 void oled_task_user(void) {
-    #if defined(KEYBOARD_splitkb_kyria_rev1)
-    render_status();
-    #elif defined(KEYBOARD_crkbd_rev1)
-    if (is_keyboard_left()) {
-        render_status_left();
-    } else {
-        render_status_right();
+    if (oled_task_keymap()) {
+        render_status();
     }
-    #endif
 }
